SeqData

SeqData is a small package that provides a data type for sequential data for learning problems. It saves a lot of memory as it does not save the identical data many times but uses a modified indexing.

Usage

Everything you need to now is in the doc string:

SeqData.SequentialData — Type

SequentialData

Struct with initialization for preparing sequence data for a learning problem. It saves a lot of memory because it uses Julia's iterator interface and does not save the same data multiple times as a naive implementation would do. The default initialization splits the data into a train, valid and test set.

SequentialData can be used for for-loops with for i in data and also indexed with data[i]. Each element is a 2-element array containing the input and output for the learning problem.

It is possible to output batches of data.

Fields of the struct

data::AbstractArray: Raw Data as a $N_dim\times N_i\times N_t$ array

N_batch::Int: Batch Size

N_length::Int: Length of each sample

N::Int: Total number of individual train/valid/test data pairs available

N_t::Int: Number of time steps per trajectory

Initialization

SequentialData(input_data::AbstractArray, N_batch::Int, N_length::Int, N_train::Int, N_valid::Int; supervised::Bool=false)

Input

input_data::AbstractArray: Raw data in either a $N_dim \times N_t \times N_i$ or $N_dim \times N_t$ array

N_batch::Int: Batch size, If N_batch==0 the output will be 2D without any batching

N_length::Int: Length of each sample

N_train::Real: Relative Number of input samples for training

N_valid::Real: Relative Number of input samples for validation. All remaining time steps will form the test set

supervised::Bool: If true every indexing will return a pair of arrays where the second array is shifted by 1 from the first array.

If $N_i > 1$ all values are refering to each trajectory seperately.

Output

Returns three instances of SequentialData in order (train_set, valid_set, test_set).